backend/requirements.txt
cached-property==1.5.2
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.9
h5py==3.1.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.4
psycopg2-binary==2.8.6
six==1.15.0
Werkzeug==1.0.1
